requirements.txt
streamlit
langchain
langchain-core
langchain-google-genai
langgraph
langsmith
python-dotenv
pyTelegramBotAPI
google-generativeai